+++ src/entrypoints/popup/main.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
+import { initTheme } from '@services/appearance/theme.ts';
 import ReactDOM from 'react-dom/client';
 import { initI18n } from '@services/i18n/i18n.ts';
 import PopupApp from './PopupApp';
 
-initI18n().then(() => {
+Promise.all([initI18n(), initTheme()]).then(() => {
   ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
       <PopupApp />